// src/app/shared/request.interface.ts
export interface Request {
    usuario: string;
    sector: string;
    categoria: string;
    subcategoria: string;
    cantidad: string;
    direccion: string;
    ciudad: string;
    fecha: string;
    pais: string;
    producto: string;
    referencia: string;
    descripcion: string;
    imagen: string;
    marca: string;
    fileRef: string;
}